Moved the symbol submodule into the sim module, as it didn't really fit in the frequency module
