+++ src/frequencies/mod.rs
@@ -17,7 +17,6 @@
 
 pub mod mutable_table;
 pub mod static_table;
-pub mod symbol;
 #[cfg(test)]
 mod unit_tests;
