+++ src/sim/symbol.rs
